TLS 1.3 Server and Encrypted Extensions Part 1
- This commit disables extensions that should not be sent in TLS 1.3
  server extensions. These should be move to their new extention
  destinations with relation to RFC8446

+++ tls/s2n_tls.h
@@ -75,7 +75,7 @@ extern int s2n_read_full_record(struct s2n_connection *conn, uint8_t * record_ty
 extern int s2n_recv_close_notify(struct s2n_connection *conn, s2n_blocked_status * blocked);
 extern int s2n_client_extensions_send(struct s2n_connection *conn, struct s2n_stuffer *out);
 extern int s2n_client_extensions_recv(struct s2n_connection *conn, struct s2n_array *parsed_extensions);
-extern uint16_t s2n_server_extensions_send_size(struct s2n_connection *conn);
+extern int s2n_server_extensions_send_size(struct s2n_connection *conn);
 extern int s2n_server_extensions_send(struct s2n_connection *conn, struct s2n_stuffer *out);
 extern int s2n_server_extensions_recv(struct s2n_connection *conn, struct s2n_blob *extensions);
